backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb>=0.6.0
scikit-learn==1.5.2
numpy>=1.26.4
pydantic==2.10.0
umap-learn==0.5.7
hdbscan==0.8.40
